Linux build fixes

+++ src/read_file_nonblocking.cc
@@ -119,13 +119,13 @@ int main(int argc, char* argv[]) {
     int remaining = 0;
     fd_set read_fds, write_fds;
     int num_fds = 0;
-    for (off_t i = 0; i < file_size; i += 1024*1024) {
+    for (int64_t i = 0; i < file_size; i += 1024*1024) {
         unsigned int block_length = 1024*1024;
         if (i + block_length > file_size) {
             block_length = file_size - i + 1;
         }
 
-        sprintf(key_buffer, "%s-%10" PRId64, kinetic_key, (int64_t)i);
+        sprintf(key_buffer, "%s-%10" PRId64, kinetic_key, i);
         remaining++;
         TestCallback* callback = new TestCallback(output_buffer + i, block_length, &remaining);
         std::string key(key_buffer);
